+++ internal/metrics/metrics.go
@@ -7,7 +7,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/dgraph-io/ristretto"
 	"github.com/inconshreveable/log15"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/sourcegraph/sourcegraph/internal/httpcli"
@@ -130,38 +129,6 @@ func MustRegisterDiskMonitor(path string) {
 		var stat syscall.Statfs_t
 		_ = syscall.Statfs(path, &stat)
 		return float64(stat.Blocks * uint64(stat.Bsize))
-	}))
-}
-
-// MustRegisterRistrettoMonitor exports three prometheus metrics
-// "src_ristretto_cache_hits{cache=$cache}",
-// "src_ristretto_cache_misses{cache=$cache}", and
-// "src_ristretto_cache_cost{cache=$cache}".
-//
-// It is safe to call this function more than once for the same cache name.
-func MustRegisterRistrettoMonitor(cacheName string, metrics *ristretto.Metrics) {
-	mustRegisterOnce(prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name:        "src_ristretto_cache_hits",
-		Help:        "Total number of cache hits.",
-		ConstLabels: prometheus.Labels{"cache": cacheName},
-	}, func() float64 {
-		return float64(metrics.Hits())
-	}))
-
-	mustRegisterOnce(prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name:        "src_ristretto_cache_misses",
-		Help:        "Total number of cache misses.",
-		ConstLabels: prometheus.Labels{"cache": cacheName},
-	}, func() float64 {
-		return float64(metrics.Misses())
-	}))
-
-	mustRegisterOnce(prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name:        "src_ristretto_cache_cost",
-		Help:        "Current cost of the cache.",
-		ConstLabels: prometheus.Labels{"cache": cacheName},
-	}, func() float64 {
-		return float64(metrics.CostAdded() - metrics.CostEvicted())
 	}))
 }
